backend/rag-service/requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# LLM & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
